# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Icarus API
  description: Contract v1 for the Icarus server (shared/api-contract.md). Times are RFC 3339 UTC with Z, except fields ending in _ms (epoch milliseconds). Metrics are estimates, not medical readings.
  version: 1.0.0
paths:
...
  /v1/devices:
    get:
      tags:
      - Devices
      operationId: list_devices
      responses:
        '200':
          description: Devices and bands of the account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceList'
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - session: []
      - device: []
  /v1/devices/me/push-token:
    put:
      tags:
      - Devices
      operationId: put_push_token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushTokenBody'
        required: true
      responses:
        '204':
          description: Stored for this device.
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - device: []
  /v1/devices/pair:
    post:
      tags:
      - Devices
      summary: Exchanges a single-use pairing code for a device token. The token is returned once.
      operationId: pair_device
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairBody'
        required: true
      responses:
        '201':
          description: Paired. The token is shown once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PairResponse'
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/devices/pairing-codes:
    post:
      tags:
      - Devices
      operationId: create_pairing_code
      responses:
        '200':
          description: A single-use code, valid for 10 minutes, and its QR code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PairingCodeResponse'
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - session: []
  /v1/devices/{id}:
    delete:
      tags:
      - Devices
      summary: Revokes a device. Its token gets 401 from the next request on.
      operationId: revoke_device
      parameters:
      - name: id
        in: path
        description: Id of the resource.
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Device revoked. Its token gets 401 from now on.
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - session: []
...
components:
  schemas:
...
    Band:
      type: object
      required:
      - id
      - created_at
      properties:
        created_at:
          type: string
          format: date-time
        firmware:
          type:
          - string
          - 'null'
        id:
          type: string
          format: uuid
        last_seen_at:
          type:
          - string
          - 'null'
          format: date-time
        name:
          type:
          - string
          - 'null'
...
    Device:
      type: object
      required:
      - id
      - name
      - created_at
      properties:
        app_version:
          type:
          - string
          - 'null'
        created_at:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        last_seen_at:
          type:
          - string
          - 'null'
          format: date-time
        model:
          type:
          - string
          - 'null'
        name:
          type: string
        os_version:
          type:
          - string
          - 'null'
        revoked_at:
          type:
          - string
          - 'null'
          format: date-time
    DeviceList:
      type: object
      required:
      - devices
      - bands
      properties:
        bands:
          type: array
          items:
            $ref: '#/components/schemas/Band'
        devices:
          type: array
          items:
            $ref: '#/components/schemas/Device'
...
    PairBody:
      type: object
      required:
      - code
      - name
      properties:
        app_version:
          type:
          - string
          - 'null'
        code:
          type: string
        model:
          type:
          - string
          - 'null'
        name:
          type: string
        os_version:
          type:
          - string
          - 'null'
    PairResponse:
      type: object
      required:
      - device_id
      - token
      properties:
        device_id:
          type: string
          format: uuid
        token:
          type: string
    PairingCodeResponse:
      type: object
      required:
      - code
      - qr_svg
      - expires_at
      properties:
        code:
          type: string
        expires_at:
          type: string
          format: date-time
        qr_svg:
          type: string
    Problem:
      type: object
      description: RFC 9457 problem document. Handlers build it in `error.rs`; this type only documents it.
      required:
      - type
      - title
      - status
      - detail
      properties:
        current:
          type:
          - object
          - 'null'
          description: The stored entity, on `conflict` responses only.
        detail:
          type: string
        status:
          type: integer
          format: int32
          minimum: 0
        title:
          type: string
        type:
          type: string
          description: '`urn:icarus:problem:<slug>`, where the slug is one of the contract''s list.'
    PushTokenBody:
      type: object
      required:
      - apns_token
      - environment
      properties:
        apns_token:
          type: string
        environment:
          type: string
...
  securitySchemes:
    device:
      type: http
      scheme: bearer
    session:
      type: apiKey
      in: cookie
      name: icarus_session
      description: 'Website session from POST /v1/auth/login. Non-GET requests also send X-Icarus-CSRF: 1.'
...